Move `ProveRpc` into its own module
